+++ reactApp/customMaps/customStyleMap.js
