+++ reactApp/labels/colorLabels.js
